+++ CHANGELOG.md
@@ -9,6 +9,7 @@ O versionamento deste projeto é aderente aos princípios de [Semantic Versionin
 ### Added
 
 - Criação do componente `<cps-tab-item>`.
+- Criação do componente `<cps-tab-panel>`.
 
 ## 0.8.1 - 2023-08-25